Move VersionCommand back to influxd
This commit moves the command in cmd/influxd/version into the
cmd/influxd package so that the version and commit can be injected
via ldflags.
